add comments to models and introduce new properties organize routes separate user controllers from auth
