+++ models/chat.js
@@ -1,6 +1,8 @@
 import mongoose from "mongoose";
 const Schema = mongoose.Schema;
 
+// Resembles a message more than a chat
+// To be refactored by Jason
 const Chat = new Schema(
     {
         sender: { type: Schema.Types.ObjectId, ref: 'User' },
