+++ routes/project/projects.js
@@ -10,31 +10,20 @@ import {
 } from '../controllers/projects.js';
 
 const router = Router();
-// standard crud
-// when would the frontend need to getAllProjects?
+
+// TODO: when would the frontend need to getAllProjects?
 router.get('/projects', getAllProjects);
 router.get('/projects/:id', getOneProject);
-// This route should be protected by BC ADMIN
+// TODO: This route should be protected by BC ADMIN
 router.post('/projects', createProject);
 router.get('/users/projects/:userId', getUserProjects);
-// what elements of a project can be updated?
-// should they each have their own route? aka, meetings, chats, tasks etc.
+// TODO: Discuss during SWE meeting
+  // What elements of a project can be updated?
+  // Should they each have their own route? aka, meetings, chats, tasks etc.
+    // Yes IMO if it's a change to only one property to protect the rest of the 
+    // record in case additional accidental data is submitted with this request
 router.put('/projects/:id', updateProject); //update description etc
 router.delete('/projects/:id', deleteProject);
-
-// special endpoints
 router.patch('/update-user-and-project', updateUserAndProject);
-// router.put(
-//   "/projects/:projectId/remove-member",
-//   controllers.removeMemberFromProject
-// );
-// router.put(
-//   "/projects/:projectId/add-interested-user",
-//   controllers.addInterestedUser
-// );
-// router.put(
-//   "/projects/:projectId/remove-interested-user",
-//   controllers.removeInterestedUser
-// );
 
 export default router;
